requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-genai==0.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
tenacity==8.2.3
gunicorn==21.2.0